fixed bug

+++ src/controllers/Admin/customers.controller.js
@@ -169,11 +169,13 @@ export const deleteCustomer = async (req, res, next) => {
 
       const user = await Users.findOne({ where: { id: customerId }, transaction });
 
-      await Otps.destroy({
-        where: { phoneCountryCode: user.phoneCountryCode, phoneNumber: user.phoneNumber },
-        force: true,
-        transaction
-      });
+      if (user.phoneCountryCode && user.phoneNumber) {
+        await Otps.destroy({
+          where: { phoneCountryCode: user.phoneCountryCode, phoneNumber: user.phoneNumber },
+          force: true,
+          transaction
+        });
+      }
 
       await user.destroy({ force: true, transaction });
     });
